refactor: 사용 안하는 코드 정리, 클래스 패키지 이동
