+++ back/src/main/java/com/ll/goohaeyou/global/standard/base/KwType.java
@@ -1,6 +1,8 @@
 package com.ll.goohaeyou.global.standard.base;
 
 import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
 
 @AllArgsConstructor
 public enum KwType {
